backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# NASA API Integration
aiohttp==3.9.1
httpx==0.25.1

# Database (optional for now)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication (optional for now)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
slowapi==0.1.8
pydantic-settings==2.1.0
psutil==5.9.6

# Enhanced Features
websockets==12.0  # For WebSocket support
pydantic==2.5.0   # Data validation
click==8.1.7      # CLI support
colorama==0.4.6   # Terminal colors

# Monitoring & Logging
prometheus-client==0.19.0  # Metrics
structlog==23.2.0          # Structured logging

# Development (optional)
black==23.11.0    # Code formatting
isort==5.12.0     # Import sorting
pytest==7.4.3     # Testing
pytest-asyncio==0.21.1  # Async testing